Added the problems & solutions for Week-1 home tasks.

+++ Week-01/ClassTask/Task05.java
@@ -1,13 +1,13 @@
 import java.util.Scanner;
 
 public class Task05 {
-    public static void main(String [] args) {
+    public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
         System.out.print("Give a value for inch: ");
-        int inch = sc.nextDouble();
+        int inch = sc.nextInt();
 
-        System.out.println(inch + " inch is " + inch*0.0254 + " meters");
+        System.out.println(inch + " inch is " + inch * 0.0254 + " meters");
         sc.close();
     }
 }
